Remove pipeline deploy path filters
Removing this list of allowed paths to trigger the pipeline to come in
line with the other repos.

Unless this list of paths is fiercly kept accurate/up-to-date then it
inevitably leads to someone being caught off guard and a pipeline not
being kicked off when it's required. Crucial changes already out of date
are the requirements files, and not redeploying after any changes to
copilot manifests.