FS-4917 - Remove unused self-serve functionality
This self-serve system was a prototype for constructing forms from lower-level components, built on top of an in-memory "database". Such low-level form construction is not currently within the scope of FAB as a production asset.

We also tidy up usage of the helper functions human_to_kebab_case and human_to_snake_case, which were previously duplicated and imported from the routes file in the self-serve directory.