FLS-1194 - Replace internal user check with allowed domains check
This change replaces the check on internal user with a FAB-specific check on allowed domains, such that we can provide access to FAB not only to users with communities.gov.uk email addresses, but also email addresses with other domains, configured via AWS secrets to allow for environment-level configuration. We are specifically doing this now because we have an accessibility audit coming up and need to ensure assessors can access.