FLS-1128 - Fixing flaky E2E tests
The fundamental issue was that a random application was being selected each time on the 'Select application' page, because the available options were not being matched properly. This caused downstream effects e.g., if the application selected was marked as 'Complete' already, then instead of going to the 'Build application' page you'd go to the more limited 'View application' page, meaning not only would the check for the text 'Build application' fail, but all of the functionality under test (e.g., moving sections up and down) wouldn't be available.

The solution is to grab the round_id from the application success banner on the dashboard page, so that we can select an application based off of the round_id value. This means we always end up with the application that we created in the E2E test rather than a random one, and this means that it'll always be 'In progress' rather than 'Complete'.

We've also simplified the way in which grants are selected, leveraging the existing grant_id metadatum.