FSPT-342 - Preview forms using Form Runner in Test env not UAT
We are wanting to remove the UAT environment as it is an unnecessary expense. We therefore need to decouple FAB in production from Form Runner in UAT so that when we remove UAT, FAB in production won't be affected in any way. Prod FAB is tied to UAT Runner because Prod Runner does not allow requests to the /publish endpoint required for previewing forms.

This is intended as a temporary measure to unblock UAT environment shutdown. Longer-term we will look to link Prod FAB to Prod Runner.