FSPT-342 - Support Form Runner basic auth
Form Runner is having basic auth added to it, as a compromise given we need to remove JWT auth from certain Runner endpoints to allow for cross-environment form publishing (Prod FAB -> Test Runner). Basic auth will create a dialogue box for credential entry from the user if the request to Runner originates from the browser, but in the case of backend requests (e.g., to /publish), we need to handle basic auth on the backend. The way we have chosen to implement this here is in line with existing precedents in Pre-Award and Post-Award - we hard-code the basic auth credentials in a URL we store in AWS Parameter Store (e.g., http://username:password@example.com) and then import that variable as a secret to swap directly into existing application code, with minimal changes required.
We have chosen to use FORM_RUNNER_PUBLISH_URL from AWS because FORM_RUNNER_INTERNAL_HOST cannot be shared with Pre-Award - this is because in the AWS prod environment, Pre-Award needs to get rehydration tokens from Prod Runner, while FAB needs to publish to Test Runner. Plus FORM_RUNNER_PUBLISH_URL already exists for use by Form Designer anyway.