FLS-1401 - Adding descriptive labels to links
This is to help visually impaired users get an unambiguous understanding via their screen reader of which entity a link's action is referring to. We specifically looked for links on pages where the same link text is used again and again to support performing a particular action against multiple entities.

Unit tests needed to be updated as well because we can no longer use soup.find- the visually hidden text spans inside the <a> tags mean the tags' string attribute is now set to None, as per https://www.crummy.com/software/BeautifulSoup/bs4/doc/#string.