FLS-1408 - Prevent start page controller being assigned during form import
Currently FAB assigns the 'controller' attribute 'start.js' to any page in a form whose path matches the form's startPage. Having this attribute triggers the Form Runner to use the StartPageController to render the page, which means we lose things like error handling as start pages are not expected by the Runner to contain input components, rather introductory spiel only. And if the unsuspecting content designer fails to include a start page in a form, then because the first page in a form will be automatically assigned as the start page by the Designer, then on import FAB will add the 'controller' attribute and destroy error handling for that page.

The automatic addition of the 'controller' attribute during form import by FAB seems to serve no purpose whatsoever, especially because during export we create a new start page automatically if it's missing. It looks like a bug that's been kicking around at the heart of FAB for far too long, and now it's time to squash it.