FLS-1408 - Don't show 'Check your answers' on auto-generated start pages
This is relevant now because previously, it was not possible for auto-generation of start pages to actually get hit. This was because if no start page existed on form import, FAB would add the 'controller' attribute 'start.js' to the first page. Then when we export the same form, FAB would always identify the start page by this attribute, and the auto-generation of start page would be skipped.

Now however, we have removed the automatic assignment of the 'controller' attribute during form import. This is because it is a straight-up bug that basically removes the error-handling capability of the Form Runner on the first pages of forms if the content designer forgets to add a start page to the form - so not great stuff. But the unlocking of start page auto-generation means we need to ensure it is fit for purpose. This change removes the redundant 'Check your answers' from the list of questions we tell the user we're about to ask them on the start page.