FLS-1468 - Stop cloning relational entities on template clone
When we add a form to an application tasklist, we clone the “template” form to create a new one tied to the funding round being configured. As part of this process, we’ve historically done a clone cascade, meaning that we’ve cloned all related entities in the forms relational milieu - Forms, FormSections, Lizts, Conditions, Pages and Components.

This commit removes all of those parts of the clone cascade that are no longer relevant, drastically simplifying the clone.py module.