FLS-1470 - Removing test_integration module as no longer relevant
Why is it no longer relevant? Because it was built to test that forms weren't corrupted between import and export, which made sense given the complexity of the de-construction / re-construction logic. However, given that we are now simply storing the form JSON as-is, validating this fact in painstaking detail feels like extra test code we just don't need.