FLS-1470 - Fixing up edit_template endpoint
Now that we are moving from storing forms relationally to as whole JSON objects, we no longer need the current convoluted upload process, whereby if we get a new form JSON uploaded, we delete the existing form and import an entirely new one. Instead, we can update the form in-place using the existing update_form interface - we just need to pass form_json to it.