Bugfix - strip leading numbers
This function was faulty. If for example you had a section name 'Section 1', it would iterate over all of the characters, see the '1' at the end, and then lop the first character from the start of the string (the 'S'), giving 'ection 1'. Not ideal!

The consequences of this function being broken? It was only used in 'view questions' HTML pages, and from inspecting the data in production, we can see that we've fortunately not had any section names that have included digits or full stops after the initial prefix, so this will have had no effect on users.