FLS-1453 - Insert new forms during round cloning instead of cloning forms
Round cloning happens when a user clicks "Copy application". Historically, we take all of the forms associated with the application and clone them. Now however, the concept of forms as templates and cloning forms doesn't exist - forms in FAB are simply links between a FAB application and forms stored in the Pre-Award database. In this change we reflect that reality in the code, using our new insert_new_section_form function instead of clone_single_form when doing a round clone.