FLS-1453 - Remove extraneous fields from Form model
The fields is_template, template_name and source_template_id are all irrelevant in a world where FAB doesn't manage forms as "templates" internally. Forms in FAB are now just links between FAB applications and forms stored in the Pre-Award database. Regarding audit_info, this has never been used in application code and was always bound for the chop one day - now felt like a good time to do it, seeing as this is a commit specifically focused on removing unnecessary columns from the form table.