FLS-1453 - Remove name_in_apply_json from Form model
This field provided the form "display name" - the thing used to represent forms in tasklists in Apply and also Assess (where forms are directly mapped to subcriteria). We don't want FAB to manage this itself - this creates duplicate sources of truth. All FAB needs is an ID, in this case the url_path that links it to the form store in the Pre-Award database. If FAB needs a display name for a form, which it does in several cases, it can go ask Pre-Award for it. It does this in two different ways - where we need display names for multiple forms (e.g., the "Build application" page), we get all published forms via the GET /forms endpoint, create a URL path -> display name mapping, and index into it; and where we only need the display name of a single form (e.g., at the view_form_questions endpoint), we use the GET /forms/{id}/published endpoint.