FLS-1453 - Remove name_in_apply_json from Form model
This field provided the form "display name" - the thing used to represent forms in tasklists in Apply and also Assess (where forms are directly mapped to subcriteria). We don't want FAB to manage this itself - this creates duplicate sources of truth. All FAB needs is an ID, in this case the url_path that links it to the form store in the Pre-Award database. If FAB needs a display name for a form, which it does in several cases, it can go ask Pre-Award for it. It does this via the GET /forms endpoint, which returns a list of published forms, each with their URL path and display name. Then FAB can create a URL path -> display name mapping form this response, and index into it.