FLS-1453 - Add universal form store API service method for getting display name from URL path to reduce duplicate code
In the last commit we started getting form display names from the Pre-Award (form store) API, in most cases using the GET /forms API endpoint. Using this directly led to duplicate code in many places in the FAB codebase, because in each instance to get a display name from a URL path we had to create a URL path -> display name mapping from the API response, and then index into it on the form of interest's URL path. This commit adds a helper method to the API service module to reduce code duplication and promote consistency. We make the same number of network requests as before by using a request-scoped caching mechanism.