FLS-1453 - Replace insert_new_form with insert_new_section_form in tests and remove insert_new_form
insert_new_section_form is the replacement for insert_new_form, offering a new, simplified interface compatible with the planned final state of FAB's form table. Usage of insert_new_form has been gradually phased out of application code over the past couple of commits, and now it's time to get rid of the function completely. The tricky part is the fact that insert_new_form is used in unit tests, so we need to replace its usage there with usage of insert_new_section_form. insert_new_section_form requires a section ID, so we pass a new section fixture. And section requires a round ID, so we pass this as a fixture too. And to get round ID, we need the seed test data, so we compile all of the database management fixtures into this top-level fixture. Doing this has the benefit of meaning we centralise our usage of fixtures and can greatly simplify a number of tests throughout the test_application module, which we duly do here.