FLS-1453 - Remove runner_publish_name from form table
Next column to be removed is runner_publish_name. This is a simple one to remove, as it is analogous to the column we created earlier - url_path. The only thing we need to be careful of at this stage is ensuring we don't lose any data. Thus we have added a data migration into the Alembic schema migration, to transfer data from the runner_publish_name column into the url_path column at the exact point in time we delete this column, ensuring continuity of service in deployed environments whithout having to worry about syncing up deployments and script runs.