FLS-1453 - Allow PR-scoped E2E tests to pass with a stub server
The FAB repo has a cool feature whereby it can run E2E tests within PRs, so before merge, not just after merge. It does this by spinning up the service as a Docker container in a special GitHub workflow, and then executing E2E tests against this container. However, we have now introduced a dependency on the Pre-Award API, and given the isolation of our FAB Docker container, this has broken E2E tests. The solution is to run a stub server representing the Pre-Award service alongside the FAB service as part of our special PR GitHub workflow. The stub server is a Flask app served on port 8001, and by overwriting the database URL environment variable with localhost:8001 when initialising the FAB service, API requests are redirected to the stub server, which sends back the data the E2E tests expect.